requirements.txt
pdf2image==1.16.0
Pillow==9.2.0
PySimpleGUI==4.60.1
